Fix  'sign compare' gcc warnings

+++ firmware/controllers/can/can_bench_test.cpp
@@ -97,7 +97,7 @@ static void qcSetEtbState(uint8_t dcIndex, uint8_t direction) {
 }
 
 static void setPin(const CANRxFrame& frame, [[maybe_unused]] int value) {
-		size_t outputIndex = frame.data8[2];
+		int outputIndex = frame.data8[2];
 		if (outputIndex >= getBoardMetaOutputsCount()) {
 		  criticalError("QC pin index %d out of range", outputIndex);
 			return;
